app/model/user_type.py
from create_db import db

class UserType(db.Model):
    __tablename__ = 'user_types'
    id = db.Column(db.Integer, primary_key=True)
    name = db.Column(db.String(30), nullable=False)

#tipos criados:
#ADMIN
#DEFAULT